+++ girderformindlogger/api/api_main.py
@@ -3,8 +3,8 @@ import cherrypy
 
 from . import describe
 from .v1 import activity, activity_set, api_key, applet, assetstore, file,     \
-    collection, context, folder, group, item, resource, response, screen,      \
-    system, token, user, notification
+    collection, context, folder, group, item, resource, response, schedule,    \
+    screen, system, token, user, notification
 
 
 class ApiDocs(object):
@@ -40,6 +40,7 @@ def _addV1ToNode(node):
     node.v1.notification = notification.Notification()
     node.v1.resource = resource.Resource()
     node.v1.response = response.ResponseItem()
+    node.v1.schedule = schedule.Schedule()
     node.v1.screen = screen.Screen()
     node.v1.system = system.System()
     node.v1.token = token.Token()